collapsing asb smoke tests to single project
