--- src/ServiceControl.Monitoring.SmokeTests.ASB.EndpointOrientedTopology/DefaultServer.cs
@@ -1,119 +1,0 @@
-﻿namespace ServiceControl.Monitoring.SmokeTests.ASB.EndpointOrientedTopology
-{
-    using NServiceBus;
-    using NServiceBus.AcceptanceTesting.Customization;
-    using NServiceBus.AcceptanceTesting.Support;
-    using NServiceBus.Hosting.Helpers;
-    using NServiceBus.ObjectBuilder;
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Reflection;
-    using System.Threading.Tasks;
-
-    public class DefaultServer : IEndpointSetupTemplate
-    {
-        public static string ConnectionString => GetEnvironmentVariable($"{nameof(AzureServiceBusTransport)}.ConnectionString");
-
-        public Task<EndpointConfiguration> GetConfiguration(RunDescriptor runDescriptor, EndpointCustomizationConfiguration endpointConfiguration, Action<EndpointConfiguration> configurationBuilderCustomization)
-        {
-            var builder = new EndpointConfiguration(endpointConfiguration.EndpointName);
-            var types = GetTypesScopedByTestClass(endpointConfiguration);
-
-            builder.TypesToIncludeInScan(types);
-
-            var transportConfig = builder.UseTransport<AzureServiceBusTransport>();
-            transportConfig.ConnectionString(ConnectionString);
-
-            var endpointOrientedTopology = transportConfig.UseEndpointOrientedTopology();
-            foreach (var publisher in endpointConfiguration.PublisherMetadata.Publishers)
-            {
-                foreach (var eventType in publisher.Events)
-                {
-                    endpointOrientedTopology.RegisterPublisher(eventType, publisher.PublisherName);
-                }
-            }
-
-            builder.UseSerialization<NewtonsoftSerializer>();
-            builder.EnableInstallers();
-
-            transportConfig.Sanitization()
-                .UseStrategy<ValidateAndHashIfNeeded>();
-
-            // w/o retries ASB will move attempted messages to the error queue right away, which will cause false failure.
-            // ScenarioRunner.PerformScenarios() verifies by default no messages are moved into error queue. If it finds any, it fails the test.
-            builder.Recoverability().Immediate(retriesSettings => retriesSettings.NumberOfRetries(3));
-
-            builder.RegisterComponents(r => { RegisterInheritanceHierarchyOfContextOnContainer(runDescriptor, r); });
-
-            configurationBuilderCustomization(builder);
-
-
-            return Task.FromResult(builder);
-        }
-
-        static void RegisterInheritanceHierarchyOfContextOnContainer(RunDescriptor runDescriptor, IConfigureComponents r)
-        {
-            var type = runDescriptor.ScenarioContext.GetType();
-            while (type != typeof(object))
-            {
-                r.RegisterSingleton(type, runDescriptor.ScenarioContext);
-                type = type.BaseType;
-            }
-        }
-
-        static IEnumerable<Type> GetTypesScopedByTestClass(EndpointCustomizationConfiguration endpointConfiguration)
-        {
-            var assemblies = new AssemblyScanner().GetScannableAssemblies();
-
-            var types = assemblies.Assemblies
-                //exclude all test types by default
-                .Where(a =>
-                {
-                    var references = a.GetReferencedAssemblies();
-
-                    return references.All(an => an.Name != "nunit.framework");
-                })
-                .SelectMany(a => a.GetTypes());
-
-
-            types = types.Union(GetNestedTypeRecursive(endpointConfiguration.BuilderType.DeclaringType, endpointConfiguration.BuilderType));
-
-            types = types.Union(endpointConfiguration.TypesToInclude);
-
-            return types.Where(t => !endpointConfiguration.TypesToExclude.Contains(t)).ToList();
-        }
-
-        static IEnumerable<Type> GetNestedTypeRecursive(Type rootType, Type builderType)
-        {
-            if (rootType == null)
-            {
-                throw new InvalidOperationException("Make sure you nest the endpoint infrastructure inside the TestFixture as nested classes");
-            }
-
-            yield return rootType;
-
-            if (typeof(IEndpointConfigurationFactory).IsAssignableFrom(rootType) && rootType != builderType)
-            {
-                yield break;
-            }
-
-            foreach (var nestedType in rootType.GetNestedTypes(BindingFlags.Instance | BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic).SelectMany(t => GetNestedTypeRecursive(t, builderType)))
-            {
-                yield return nestedType;
-            }
-        }
-
-        public static string GetEnvironmentVariable(string variable)
-        {
-            var candidate = Environment.GetEnvironmentVariable(variable, EnvironmentVariableTarget.User);
-
-            if (string.IsNullOrWhiteSpace(candidate))
-            {
-                return Environment.GetEnvironmentVariable(variable);
-            }
-
-            return candidate;
-        }
-    }
-}
