--- src/ServiceControl.Monitoring.SmokeTests.ASB.ForwardingTopology/packages.config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net46" />
-  <package id="NServiceBus" version="7.1.4" targetFramework="net46" />
-  <package id="NServiceBus.AcceptanceTesting" version="7.1.4" targetFramework="net46" />
-  <package id="NServiceBus.Azure.Transports.WindowsAzureServiceBus" version="8.0.3" targetFramework="net46" />
-  <package id="NServiceBus.Metrics" version="3.0.0" targetFramework="net46" />
-  <package id="NServiceBus.Metrics.ServiceControl" version="3.0.2" targetFramework="net46" />
-  <package id="NServiceBus.Newtonsoft.Json" version="2.1.0" targetFramework="net46" />
-  <package id="NUnit" version="3.10.1" targetFramework="net46" />
-  <package id="WindowsAzure.ServiceBus" version="4.1.3" targetFramework="net46" />
-</packages>
