+++ src/ServiceControl.Monitoring.SmokeTests.ASB/Tests/When_querying_retries_data.cs
@@ -1,31 +1,35 @@
-﻿namespace ServiceControl.Monitoring.SmokeTests.ASB.ForwardingTopology.Tests
+﻿namespace ServiceControl.Monitoring.SmokeTests.ASB.Tests
 {
+    using System;
+    using System.Threading.Tasks;
     using Newtonsoft.Json.Linq;
     using NServiceBus;
     using NServiceBus.AcceptanceTesting;
     using NUnit.Framework;
-    using System;
-    using System.Threading.Tasks;
-    using Conventions = NServiceBus.AcceptanceTesting.Customization;
 
     [Category("TransportSmokeTests")]
-    public class When_querying_timings_data : ApiIntegrationTest
+    public class When_querying_retries_data : ApiIntegrationTest
     {
-        static string ReceiverEndpointName => Conventions.Conventions.EndpointNamingConvention(typeof(MonitoringEndpoint));
+        static string ReceiverEndpointName => NServiceBus.AcceptanceTesting.Customization.Conventions.EndpointNamingConvention(typeof(MonitoringEndpoint));
 
         [Test]
         public async Task Should_report_via_http()
         {
-            JToken processingTime = null;
+            JToken retries = null;
 
             await Scenario.Define<Context>()
-                    .WithEndpoint<MonitoredEndpoint>(c => c.When(s => s.SendLocal(new SampleMessage())))
-                    .WithEndpoint<MonitoringEndpoint>()
-                    .Done(c => MetricReported("processingTime", out processingTime, c))
-                    .Run();
+                .WithEndpoint<MonitoredEndpoint>(c =>
+                {
+                    c.DoNotFailOnErrorMessages();
+                    c.CustomConfig(ec => ec.Recoverability().Immediate(i => i.NumberOfRetries(5)));
+                    c.When(s => s.SendLocal(new SampleMessage()));
+                })
+                .WithEndpoint<MonitoringEndpoint>()
+                .Done(c => MetricReported("retries", out retries, c))
+                .Run();
 
-            Assert.IsTrue(processingTime["average"].Value<int>() > 0);
-            Assert.AreEqual(60, processingTime["points"].Value<JArray>().Count);
+            Assert.IsTrue(retries["average"].Value<double>() > 0);
+            Assert.AreEqual(60, retries["points"].Value<JArray>().Count);
         }
 
         class MonitoredEndpoint : EndpointConfigurationBuilder
@@ -34,17 +38,15 @@
             {
                 EndpointSetup<DefaultServer>(c =>
                 {
-#pragma warning disable 618
                     c.EnableMetrics().SendMetricDataToServiceControl(ReceiverEndpointName, TimeSpan.FromSeconds(5));
-#pragma warning restore 618
                 });
             }
 
-            class Handler : IHandleMessages<SampleMessage>
+            public class Handler : IHandleMessages<SampleMessage>
             {
                 public Task Handle(SampleMessage message, IMessageHandlerContext context)
                 {
-                    return Task.Delay(TimeSpan.FromMilliseconds(10));
+                    throw new Exception("Boom!");
                 }
             }
         }
