+++ src/ServiceControl.Monitoring.SmokeTests.ASB/app.config
