+++ src/ServiceControl.Monitoring.SmokeTests.ASB/packages.config
