+++ src/ServiceControl.Monitoring.sln
@@ -41,8 +41,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.Monitoring.S
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SmokeTests", "SmokeTests", "{81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.Monitoring.SmokeTests.ASB.ForwardingTopology", "ServiceControl.Monitoring.SmokeTests.ASB.ForwardingTopology\ServiceControl.Monitoring.SmokeTests.ASB.ForwardingTopology.csproj", "{10B6F344-E306-41BB-AD88-8235BD1F3092}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.Monitoring.SmokeTests.Learning", "ServiceControl.Monitoring.SmokeTests.Learning\ServiceControl.Monitoring.SmokeTests.Learning.csproj", "{95D63BE5-D5EF-46A5-B38B-F68105062EDB}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.Monitoring.SmokeTests.SQLServer", "ServiceControl.Monitoring.SmokeTests.SQLServer\ServiceControl.Monitoring.SmokeTests.SQLServer.csproj", "{F98424DA-ECA8-4636-9062-93E058E2C304}"
@@ -67,11 +65,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Packaging.NServiceBus.Azure
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.Monitoring.SmokeTests.ASBS", "ServiceControl.Monitoring.SmokeTests.ASBS\ServiceControl.Monitoring.SmokeTests.ASBS.csproj", "{A82A5634-FC28-47FF-BDAA-69ABBE74A6BC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.Monitoring.SmokeTests.ASB.EndpointOrientedTopology", "ServiceControl.Monitoring.SmokeTests.ASB.EndpointOrientedTopology\ServiceControl.Monitoring.SmokeTests.ASB.EndpointOrientedTopology.csproj", "{574411CC-AFB5-4525-BF98-588D90BFF411}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceControl.Transports.AzureServiceBus", "ServiceControl.Transports.AzureServiceBus\ServiceControl.Transports.AzureServiceBus.csproj", "{A5A18CE5-2065-4E9A-A8A3-E55CABB50407}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Packaging.NServiceBus.AzureServiceBus", "Packaging.NServiceBus.AzureServiceBus\Packaging.NServiceBus.AzureServiceBus.csproj", "{B5EBF681-E632-4783-B003-08E9BB320621}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.Monitoring.SmokeTests.ASB", "ServiceControl.Monitoring.SmokeTests.ASB\ServiceControl.Monitoring.SmokeTests.ASB.csproj", "{574411CC-AFB5-4525-BF98-588D90BFF411}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -127,10 +125,6 @@ Global
 		{36DD5C1F-D189-4BF3-989D-AF454FA68226}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{36DD5C1F-D189-4BF3-989D-AF454FA68226}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{36DD5C1F-D189-4BF3-989D-AF454FA68226}.Release|Any CPU.Build.0 = Release|Any CPU
-		{10B6F344-E306-41BB-AD88-8235BD1F3092}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{10B6F344-E306-41BB-AD88-8235BD1F3092}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{10B6F344-E306-41BB-AD88-8235BD1F3092}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{10B6F344-E306-41BB-AD88-8235BD1F3092}.Release|Any CPU.Build.0 = Release|Any CPU
 		{95D63BE5-D5EF-46A5-B38B-F68105062EDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{95D63BE5-D5EF-46A5-B38B-F68105062EDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{95D63BE5-D5EF-46A5-B38B-F68105062EDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -179,10 +173,6 @@ Global
 		{A82A5634-FC28-47FF-BDAA-69ABBE74A6BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A82A5634-FC28-47FF-BDAA-69ABBE74A6BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A82A5634-FC28-47FF-BDAA-69ABBE74A6BC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{574411CC-AFB5-4525-BF98-588D90BFF411}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{574411CC-AFB5-4525-BF98-588D90BFF411}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{574411CC-AFB5-4525-BF98-588D90BFF411}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{574411CC-AFB5-4525-BF98-588D90BFF411}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A5A18CE5-2065-4E9A-A8A3-E55CABB50407}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A5A18CE5-2065-4E9A-A8A3-E55CABB50407}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A5A18CE5-2065-4E9A-A8A3-E55CABB50407}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -191,6 +181,10 @@ Global
 		{B5EBF681-E632-4783-B003-08E9BB320621}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B5EBF681-E632-4783-B003-08E9BB320621}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B5EBF681-E632-4783-B003-08E9BB320621}.Release|Any CPU.Build.0 = Release|Any CPU
+		{574411CC-AFB5-4525-BF98-588D90BFF411}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{574411CC-AFB5-4525-BF98-588D90BFF411}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{574411CC-AFB5-4525-BF98-588D90BFF411}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{574411CC-AFB5-4525-BF98-588D90BFF411}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -205,7 +199,6 @@ Global
 		{F30AC7F6-FD0A-43D2-BC75-889F9A7D2872} = {AC16BCB1-A7CA-4E50-B6BD-11518421382E}
 		{B5EB5F58-E30A-45D2-AC0D-88CEF2F804E0} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 		{36DD5C1F-D189-4BF3-989D-AF454FA68226} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
-		{10B6F344-E306-41BB-AD88-8235BD1F3092} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 		{95D63BE5-D5EF-46A5-B38B-F68105062EDB} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 		{F98424DA-ECA8-4636-9062-93E058E2C304} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 		{E3EAE283-6095-421D-B994-3E5A2674A6E0} = {ED6CE213-D704-4695-97AB-3E6EA4859795}
@@ -218,16 +211,16 @@ Global
 		{3CDA118C-521A-4FD1-BDAB-4059D141D6F2} = {ED6CE213-D704-4695-97AB-3E6EA4859795}
 		{ABA69FEC-470E-4638-9371-9D84B71605BD} = {AC16BCB1-A7CA-4E50-B6BD-11518421382E}
 		{A82A5634-FC28-47FF-BDAA-69ABBE74A6BC} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
-		{574411CC-AFB5-4525-BF98-588D90BFF411} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 		{A5A18CE5-2065-4E9A-A8A3-E55CABB50407} = {ED6CE213-D704-4695-97AB-3E6EA4859795}
 		{B5EBF681-E632-4783-B003-08E9BB320621} = {AC16BCB1-A7CA-4E50-B6BD-11518421382E}
+		{574411CC-AFB5-4525-BF98-588D90BFF411} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		{F98424DA-ECA8-4636-9062-93E058E2C304} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
-		{95D63BE5-D5EF-46A5-B38B-F68105062EDB} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
-		{10B6F344-E306-41BB-AD88-8235BD1F3092} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
-		{36DD5C1F-D189-4BF3-989D-AF454FA68226} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
-		{931BC9F4-C82D-4C3C-AC0C-054A91B77A06} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 		SolutionGuid = {D3D78742-599D-471F-AC9B-77C0807D9515}
+		{931BC9F4-C82D-4C3C-AC0C-054A91B77A06} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
+		{36DD5C1F-D189-4BF3-989D-AF454FA68226} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
+		{10B6F344-E306-41BB-AD88-8235BD1F3092} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
+		{95D63BE5-D5EF-46A5-B38B-F68105062EDB} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
+		{F98424DA-ECA8-4636-9062-93E058E2C304} = {81A9662E-CF9A-4C8A-9A2D-023973C7E1A3}
 	EndGlobalSection
 EndGlobal
